+++ src/pwshBedrock/Public/Invoke-ConverseAPI.ps1
@@ -253,11 +253,9 @@ function Invoke-ConverseAPI {
         [Parameter(Mandatory = $true,
             HelpMessage = 'The unique identifier of the model.')]
         [ValidateSet(
-            # 'ai21.j2-grande-instruct', # *note: not supported by Converse API
-            # 'ai21.j2-jumbo-instruct', # *note: not supported by Converse API
             'ai21.jamba-instruct-v1:0',
-            # 'ai21.j2-mid-v1', # *note: not supported by Converse API
-            # 'ai21.j2-ultra-v1', # *note: not supported by Converse API
+            'ai21.jamba-1-5-mini-v1:0',
+            'ai21.jamba-1-5-large-v1:0',
             # 'amazon.titan-image-generator-v1', # *note: not supported by Converse API
             # 'amazon.titan-image-generator-v2:0', # *note: not supported by Converse API
             'amazon.titan-text-express-v1',