+++ src/pwshBedrock/Public/Reset-ModelContext.ps1
@@ -41,11 +41,9 @@ function Reset-ModelContext {
             ParameterSetName = 'Single')]
         [ValidateSet(
             'Converse',
-            # 'ai21.j2-grande-instruct',
-            # 'ai21.j2-jumbo-instruct',
             'ai21.jamba-instruct-v1:0',
-            # 'ai21.j2-mid-v1',
-            # 'ai21.j2-ultra-v1',
+            'ai21.jamba-1-5-mini-v1:0',
+            'ai21.jamba-1-5-large-v1:0',
             'amazon.titan-image-generator-v1',
             'amazon.titan-image-generator-v2:0',
             'amazon.titan-text-express-v1',
@@ -133,7 +131,6 @@ function Reset-ModelContext {
                     $allModelIDs = ($allModelInfo | Where-Object {
                             $_.ModelId -ne 'amazon.titan-image-generator-v1' -and
                             $_.ModelId -ne 'amazon.titan-image-generator-v2:0' -and
-                            $_.ModelId -notlike 'ai21.j2*' -and
                             $_.ModelId -ne 'cohere.command-text-v14' -and
                             $_.ModelId -ne 'cohere.command-light-text-v14'
                         }).ModelID