+++ docs/CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.2.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.18.0] - **BREAKING CHANGES**
+## [0.21.0] - **BREAKING CHANGES**
 
 - Module changes:
     - stability.ai
@@ -14,6 +14,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
             - `Height` parameter now defaults to 1024 if no user input is provided.
             - `Width` parameter now defaults to 1024 if no user input is provided.
         - Added `Invoke-StabilityAIImageModel` to support new models: `stability.stable-image-core-v1:0`, `stability.stable-image-ultra-v1:0`, `stability.sd3-large-v1:0`
+    - AI21 Labs
+        - `Invoke-AI21LabsJambaModel` - added support for `ai21.jamba-1-5-mini-v1:0` and `ai21.jamba-1-5-large-v1:0`
+        - `Invoke-AI21LabsJurassic2Model` removed. All support for `ai21.j2-grande-instruct`, `ai21.j2-jumbo-instruct`, `ai21.j2-mid-v1`, `ai21.j2-ultra-v1` removed. These models are being [EOL](https://docs.aws.amazon.com/bedrock/latest/userguide/model-lifecycle.html) in Amazon Bedrock. - ***Breaking Change***
 - Build changes:
     - Updated bootstrap AWS modules from `4.1.621` to `4.1.690`